global styling of drop down menu done
