+++ lib/configs/configurations/theme/themes/light_theme.dart
@@ -35,21 +35,18 @@ ThemeData lightTheme = ThemeData(
     ),
   ),
   textButtonTheme: TextButtonThemeData(
-    style: ElevatedButton.styleFrom(
-      foregroundColor: dark,
-    )
-  ),
+      style: ElevatedButton.styleFrom(
+    foregroundColor: dark,
+  )),
   outlinedButtonTheme: OutlinedButtonThemeData(
-    style: ElevatedButton.styleFrom(
-      elevation: 2,
-      backgroundColor: light,
-      foregroundColor: shadow,
-      shadowColor: shadow,
-      minimumSize: const Size(230, 55),
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
-      side: const BorderSide(width: 0)
-    )
-  ),
+      style: ElevatedButton.styleFrom(
+          elevation: 2,
+          backgroundColor: light,
+          foregroundColor: shadow,
+          shadowColor: shadow,
+          minimumSize: const Size(230, 55),
+          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
+          side: const BorderSide(width: 0))),
   appBarTheme: const AppBarTheme(
     backgroundColor: light,
     iconTheme: IconThemeData(color: accent),
